backend/pyproject.toml
[tool.poetry]
name = "stock-trading-backend"
version = "0.1.0"
description = "FastAPI backend for stock trading platform"
authors = ["Stock Trading Team <team@example.com>"]
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = { extras = ["standard"], version = "^0.30.0" }
pydantic = "^2.7.1"
httpx = "^0.27.0"
sqlalchemy = "^2.0.29"
aiosqlite = "^0.20.0"
python-dotenv = "^1.0.1"
redis = "^5.0.4"
pandas = "^2.2.2"
cachetools = "^5.3.3"
cryptography = "^42.0.5"
grpcio = "^1.62.2"
protobuf = "^4.25.3"
structlog = "^24.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-asyncio = "^0.23.6"
httpx = {version = "^0.27.0", extras = ["cli"]}
coverage = "^7.4.4"
ruff = "^0.4.2"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
